[TinyglTF] Fixed minor warning

+++ Source/Dependencies/utils/include/tinyglTF/tiny_gltf.h
@@ -2940,7 +2940,7 @@ bool GetFileSizeInBytes(size_t *filesize_out, std::string *err,
       (*err) += "File is empty : " + filepath + "\n";
     }
     return false;
-  } else if (sz >= (std::numeric_limits<std::streamoff>::max)()) {
+  } else if (sz >= (size_t)(std::numeric_limits<std::streamoff>::max)()) {
     if (err) {
       (*err) += "Invalid file size : " + filepath + "\n";
     }
@@ -3033,7 +3033,7 @@ bool ReadWholeFile(std::vector<unsigned char> *out, std::string *err,
       (*err) += "File is empty : " + filepath + "\n";
     }
     return false;
-  } else if (sz >= (std::numeric_limits<std::streamoff>::max)()) {
+  } else if (sz >= (size_t)(std::numeric_limits<std::streamoff>::max)()) {
     if (err) {
       (*err) += "Invalid file size : " + filepath + "\n";
     }
